Added support for filtering to converter.
