+++ src/main/scala/org/renci/babel/utils/DiffReporter.scala
@@ -1,6 +1,7 @@
 package org.renci.babel.utils
 
 import com.typesafe.scalalogging.LazyLogging
+import org.renci.babel.utils.Utils.SupportsFilenameFiltering
 import org.renci.babel.utils.model.{BabelOutput, Compendium}
 import org.rogach.scallop.{ScallopOption, Subcommand}
 import zio.ZIO
@@ -17,7 +18,7 @@ import java.time.LocalDateTime
  */
 object DiffReporter extends LazyLogging {
   /** The subcommand that controlling comparing. */
-  class DiffSubcommand extends Subcommand("diff") {
+  class DiffSubcommand extends Subcommand("diff") with SupportsFilenameFiltering {
     val babelOutput: ScallopOption[File] = trailArg[File](
       descr = "The current Babel output directory",
       required = true
@@ -27,53 +28,17 @@ object DiffReporter extends LazyLogging {
     validateFileIsDirectory(babelOutput)
     validateFileIsDirectory(babelPrevOutput)
 
-    val filterIn: ScallopOption[List[String]] = opt[List[String]](descr =
-      "List of filenames to include (matched using startsWith)"
-    )
-    val filterOut: ScallopOption[List[String]] = opt[List[String]](descr =
-      "List of filenames to exclude (matched using startsWith)"
-    )
-
     val nCores: ScallopOption[Int] = opt[Int](descr = "Number of cores to use")
 
     val output: ScallopOption[File] = opt[File](descr = "Output file")
   }
 
   /**
-   * Generic method to determine whether a particular filename should be
-   * filtered in or out from the results. The algorithm we use is:
-   * 1. If any `--filtered-in` prefixes are provided, then we exclude everything
-   *    that isn't explicitly filtered in (by starting with one of those prefixes
-   *    in a case-sensitive manner).
-   * 2. Otherwise, all filenames are allowed EXCEPT those explicitly filtered out
-   *    by `--filtered-out` by starting with one of those prefixes in a
-   *    case-sensitive manner.
+   * Given two BabelOutputs, it returns a list of all compendia found in BOTH of the BabelOutputs
+   * paired together.
+   *
+   * TODO: modify this so we return every compendium found in EITHER BabelOutput.
    */
-  def filterFilename(conf: DiffSubcommand, filename: String): Boolean = {
-    val filteredIn = conf.filterIn.getOrElse(List())
-    val filteredOut = conf.filterOut.getOrElse(List())
-
-    if (filteredIn.nonEmpty) {
-      if (filteredIn.exists(filename.startsWith(_))) {
-        return true;
-      } else {
-        return false;
-      }
-    }
-
-    if (filteredOut.nonEmpty && filteredOut.exists(filename.startsWith(_))) {
-      return false;
-    }
-
-    true
-  }
-
-  /** Given two BabelOutputs, it returns a list of all compendia found in BOTH
-    * of the BabelOutputs paired together.
-    *
-    * TODO: modify this so we return every compendium found in EITHER
-    * BabelOutput.
-    */
   def retrievePairedCompendiaSummaries(
       babelOutput: BabelOutput,
       babelPrevOutput: BabelOutput
@@ -101,10 +66,10 @@ object DiffReporter extends LazyLogging {
       .fromIterable(pairedSummaries)
       .mapMPar(conf.nCores()) {
         case (
-              filename: String,
-              summary: Compendium,
-              prevSummary: Compendium
-            ) if filterFilename(conf, filename) => {
+          filename: String,
+          summary: Compendium,
+          prevSummary: Compendium
+          ) if Utils.filterFilename(conf, filename) => {
 
           for {
             // lengthComparison <- Comparer.compareLengths(filename, summary, prevSummary)
@@ -137,7 +102,7 @@ object DiffReporter extends LazyLogging {
             summary
           }
         }
-        case (filename: String, _, _) if !filterFilename(conf, filename) => {
+        case (filename: String, _, _) if !Utils.filterFilename(conf, filename) => {
           logger.info(s"Skipping ${filename}")
           ZIO.succeed("")
         }
